Changed config.
