--- src/main/java/com/example/demo/config/AppConfig.java
@@ -1,26 +1,0 @@
-package com.example.demo.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.ViewResolverRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
-import org.springframework.web.servlet.view.InternalResourceViewResolver;
-import org.springframework.web.servlet.view.JstlView;
-
-/**
- * Created by marta on 03.06.17.
- */
-
-@Configuration
-@EnableWebMvc
-@Import({ SecurityConfig.class })
-public class AppConfig extends WebMvcConfigurerAdapter {
-
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/static/**").addResourceLocations("/static/");
-    }
-}
