+++ src/main/java/com/example/demo/config/WebAppConfig.java
@@ -1,0 +1,26 @@
+package com.example.demo.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.core.Ordered;
+import org.springframework.web.servlet.config.annotation.*;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+
+/**
+ * Created by marta on 03.06.17.
+ */
+
+@Configuration
+@EnableWebMvc
+public class WebAppConfig extends WebMvcConfigurerAdapter {
+
+
+
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/login").setViewName("login");
+        registry.setOrder(Ordered.HIGHEST_PRECEDENCE);
+    }
+}
