+++ src/main/resources/static/js/AppConfig.js
@@ -4,6 +4,12 @@
 'use.strict'
 
 App.config(function ($routeProvider) {
-    $routeProvider.otherwise("/");
+    $routeProvider.when("/users/:id", {
+        controller: 'UserController',
+        templateUrl: '/templates/user/show.html'
+    }).when("/users", {
+        controller: 'UsersController',
+        templateUrl: '/js/user/show.html'
+    });
 
 })
