+++ src/main/resources/static/js/app.js
@@ -5,5 +5,5 @@
 'use strict';
 
 var App = angular.module('myApp',[
-    ''
+
 ]);
