+++ src/main/resources/templates/index.html
@@ -1,11 +1,12 @@
 <!DOCTYPE html >
-<html >
+<html xmlns="http://www.w3.org/1999/xhtml" >
 <head>
     <meta charset="UTF-8"/>
     <title>Title</title>
-    <link href="css/angular-bootstrap.css" rel="stylesheet"/>
+    <script src="js/app.js"></script>
+    <!-- bower:css -->
 
-
+    <!-- endbower -->
 
 </head>
 <body ng-app="myApp">
@@ -15,7 +16,9 @@
     <p ng-bind="name"></p>
 </div>
 
-<script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.6.4/angular.min.js"></script>
-<script src="js/app.js"></script>
+
+<!-- bower:js -->
+<script src="bower_components/angular/angular.js"></script>
+<!-- endbower -->
 </body>
 </html>
